Add validation function to the ClienteDelete.js component

+++ src/components/clientes/ClienteDelete.js
@@ -1,17 +1,23 @@
 import React from "react";
+import '../layout/button-styles.css';
 import Button from 'react-bootstrap/Button';
 import { FaRegTrashAlt, FaRegEdit } from "react-icons/fa";
 
 function ClienteDelete({id, handleRemove}){
   const remove =(e) =>{
     e.preventDefault()
-    handleRemove(id)
+
+    const confirmed = window.confirm("Tem certeza que deseja excluir este cliente?");
+    if (confirmed) {
+      handleRemove(id);
+    }
+
   }
 
   return(
         <div>
-          <Button variant="danger"  onClick={remove}> <FaRegTrashAlt/> </Button>
-          <Button variant="primary" href={`/clientes/${id}`} ><FaRegEdit/></Button>
+          <Button variant="danger"  className="button-styles" onClick={remove}> <FaRegTrashAlt/> </Button>
+          <Button variant="primary" className="button-styles" href={`/clientes/novocliente/${id}`} ><FaRegEdit/></Button>
       </div>
     )
 }
